requirement.txt
python-dotenv
openai
streamlit
PyPDF2